+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 #### Fixed
 - Bug where multiple legs of `Knockout` and `Tic-Tac-Toe` did not work correctly
 - Correctly calculate score if `Knockout` is won in 1 visit
+- Calculation of `420` scores per visit
 
 ## [2.1.0] - 2021-10-17
 #### Feature